pyproject.toml
[tool.poetry]
name = "SlotContrast"
version = "0.1.0"
description = "Temporally Consistent Object-Centric Learning by Contrasting Slots"
authors = [
  "Anna Manasyan",
  "Andrii Zadaianchuk"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10 <3.13"
numpy = "^1.23.3"
torchmetrics = "^1.2.0"
pytorch-lightning = "^1.8.3"
timm = "^0.9.7"
omegaconf = "^2.2.3"
webdataset = "^0.2.35"
einops = "^0.6.0"
scipy = "^1.10.1"
tensorflow-datasets = {version = "^4.7.0", optional=true}
tensorflow-cpu = {version = "^2.11.0", optional=true}
notebook = {version = "^6.5.2", optional=true}
matplotlib = {version = "^3.6.2", optional=true}
moviepy = "^1.0.3"
pycocotools = {version = "^2.0.6", optional=true}
requests = "^2.31.0"
av = "^12.0.0"

[tool.poetry.extras]
tensorflow = ["tensorflow-cpu", "tensorflow_datasets"]
coco = ["pycocotools"]
notebook = ["notebook", "matplotlib"]

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.12.0"
pytest = "^7.1.3"
pre-commit = "^2.20.0"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 101
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 101
skip_gitignore = true

[tool.ruff]
ignore = ["E731"]
line-length = 101
select = [
  "B",
  "E",
  "F",
  "TID",
  "W",
]

[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning"
markers = [
  "slow: Marks slow tests",
  "dataset: Marks tests with dataset dependencies"
]
addopts = " -m 'not slow'"  # Do not run those tests by default
